Node.js modules now build and run on Mac OS.

+++ src/qminer/qminer_core.h
@@ -2683,7 +2683,7 @@ public:
 	// stores
     void AddStore(const PStore& NewStore);
     int GetStores() const { return StoreH.Len(); }
-    bool IsStoreN(const uint& StoreN) const { return StoreN < StoreH.Len(); }
+    bool IsStoreN(const uint& StoreN) const { return StoreN < (uint)StoreH.Len(); }
     const TWPt<TStore> GetStoreByStoreN(const int& StoreN) const;
     bool IsStoreId(const uint& StoreId) const { return !StoreV[StoreId].Empty(); }
     const TWPt<TStore> GetStoreByStoreId(const uint& StoreId) const;
